Refactor Loading component imports to use loader.jsx across multiple pages
